pyproject.toml
[tool.poetry]
name = "twitter-monitor"
version = "0.1.0"
description = ""
authors = ["serkef <sergiokef@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
python-dotenv = "^0.11.0"
tweepy = "^3.8.0"
selenium = "^3.141.0"
pyzmq = "^18.1.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
